requirements.txt
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-environ==0.4.5
django-haystack==3.0
django-heroku==0.3.1
model-bakery==1.2.1
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2-binary==2.8.6
psycopg2==2.8.6
whitenoise==5.2.0
django-environ==0.4.5